fix tests SSCS-10865
