+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/adjourncase/AdjournCaseAboutToSubmitHandlerMainTest.java
@@ -3,11 +3,9 @@ package uk.gov.hmcts.reform.sscs.ccd.presubmit.adjourncase;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.DRAFT_ADJOURNMENT_NOTICE;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute.LIST_ASSIST;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.NO;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
 
@@ -86,57 +84,6 @@ class AdjournCaseAboutToSubmitHandlerMainTest extends AdjournCaseAboutToSubmitHa
 
         assertThat(response.getData().getAppeal().getHearingOptions().getLanguageInterpreter()).isEqualTo(YES.getValue());
         assertThat(response.getData().getAppeal().getHearingOptions().getLanguages()).isEqualTo(SPANISH);
-    }
-
-    @DisplayName("When adjournment is disabled and case is LA, then should not send any messages")
-    @Test
-    void givenFeatureFlagDisabled_thenNoMessageIsSent() {
-        ReflectionTestUtils.setField(handler, "isAdjournmentEnabled", false);
-        sscsCaseData.getSchedulingAndListingFields().setHearingRoute(LIST_ASSIST);
-
-        PreSubmitCallbackResponse<SscsCaseData> response =
-            handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-
-        verifyNoInteractions(hearingMessageHelper);
-
-        assertThat(response.getErrors()).isEmpty();
-    }
-
-    @DisplayName("When adjournment is enabled and case is LA and case cannot be listed right away "
-        + "and no directions are being made, then should send a new hearing request in hearings API")
-    @Test
-    void givenCaseCannotBeListedRightAwayAndNoDirectionsBeingMade_thenNewHearingRequestSent() {
-        PreSubmitCallbackResponse<SscsCaseData> response = cannotBeListedAndNoDirectionsGiven();
-
-        assertHearingCreatedAndAdjournmentInProgress(response);
-    }
-
-    @DisplayName("When adjournment is enabled and case is LA and case can be listed right away "
-        + "then should send a new hearing request in hearings API")
-    @Test
-    void givenCanBeListedRightAway_thenNewHearingRequestSent() {
-        PreSubmitCallbackResponse<SscsCaseData> response = canBeListed();
-
-        assertHearingCreatedAndAdjournmentInProgress(response);
-    }
-
-    private void assertHearingCreatedAndAdjournmentInProgress(PreSubmitCallbackResponse<SscsCaseData> response) {
-        verify(hearingMessageHelper, times(1))
-            .sendListAssistCreateHearingMessage(sscsCaseData.getCcdCaseId());
-
-        assertThat(response.getErrors()).isEmpty();
-        assertThat(response.getData().getAdjournment().getAdjournmentInProgress()).isEqualTo(YES);
-    }
-
-    @DisplayName("When adjournment is enabled and case is LA and case cannot be listed right away "
-        + "and directions are being made, then should not send any messages")
-    @Test
-    void givenCaseCannotBeListedRightAwayAndDirectionsAreBeingMade_thenNoMessagesSent() {
-        PreSubmitCallbackResponse<SscsCaseData> response = cannotBeListedAndDirectionsGiven();
-
-        verifyNoInteractions(hearingMessageHelper);
-
-        assertThat(response.getErrors()).isEmpty();
     }
 
     @DisplayName("When we have written an adjournment notice and excluded some panel members, add them "
