+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/adjourncase/AdjournCaseAboutToSubmitHandlerTest.java
@@ -3,11 +3,9 @@ package uk.gov.hmcts.reform.sscs.ccd.presubmit.adjourncase;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.DRAFT_ADJOURNMENT_NOTICE;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute.LIST_ASSIST;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.NO;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
 import static uk.gov.hmcts.reform.sscs.util.SyaServiceHelper.getRegionalProcessingCenter;
@@ -19,7 +17,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
-import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Address;
 import uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCaseTypeOfHearing;
@@ -102,57 +99,6 @@ class AdjournCaseAboutToSubmitHandlerTest extends AdjournCaseAboutToSubmitHandle
 
         assertThat(response.getData().getAppeal().getHearingOptions().getLanguageInterpreter()).isEqualTo(YES.getValue());
         assertThat(response.getData().getAppeal().getHearingOptions().getLanguages()).isEqualTo(SPANISH);
-    }
-
-    @DisplayName("When adjournment is disabled and case is LA, then should not send any messages")
-    @Test
-    void givenFeatureFlagDisabled_thenNoMessageIsSent() {
-        ReflectionTestUtils.setField(handler, "isAdjournmentEnabled", false);
-        sscsCaseData.getSchedulingAndListingFields().setHearingRoute(LIST_ASSIST);
-
-        PreSubmitCallbackResponse<SscsCaseData> response =
-            handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-
-        verifyNoInteractions(hearingMessageHelper);
-
-        assertThat(response.getErrors()).isEmpty();
-    }
-
-    @DisplayName("When adjournment is enabled and case is LA and case cannot be listed right away "
-        + "and no directions are being made, then should send a new hearing request in hearings API")
-    @Test
-    void givenCaseCannotBeListedRightAwayAndNoDirectionsBeingMade_thenNewHearingRequestSent() {
-        PreSubmitCallbackResponse<SscsCaseData> response = cannotBeListedAndNoDirectionsGiven();
-
-        assertHearingCreatedAndAdjournmentInProgress(response);
-    }
-
-    @DisplayName("When adjournment is enabled and case is LA and case can be listed right away "
-        + "then should send a new hearing request in hearings API")
-    @Test
-    void givenCanBeListedRightAway_thenNewHearingRequestSent() {
-        PreSubmitCallbackResponse<SscsCaseData> response = canBeListed();
-
-        assertHearingCreatedAndAdjournmentInProgress(response);
-    }
-
-    private void assertHearingCreatedAndAdjournmentInProgress(PreSubmitCallbackResponse<SscsCaseData> response) {
-        verify(hearingMessageHelper, times(1))
-            .sendListAssistCreateHearingMessage(sscsCaseData.getCcdCaseId());
-
-        assertThat(response.getErrors()).isEmpty();
-        assertThat(response.getData().getAdjournment().getAdjournmentInProgress()).isEqualTo(YES);
-    }
-
-    @DisplayName("When adjournment is enabled and case is LA and case cannot be listed right away "
-        + "and directions are being made, then should not send any messages")
-    @Test
-    void givenCaseCannotBeListedRightAwayAndDirectionsAreBeingMade_thenNoMessagesSent() {
-        PreSubmitCallbackResponse<SscsCaseData> response = cannotBeListedAndDirectionsGiven();
-
-        verifyNoInteractions(hearingMessageHelper);
-
-        assertThat(response.getErrors()).isEmpty();
     }
 
     @DisplayName("When we have changed the next hearing venue through an adjournment, show we change the region")
