+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/adjourncase/AdjournCaseAboutToSubmitHandlerTestBase.java
@@ -1,10 +1,5 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.adjourncase;
 
-import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute.LIST_ASSIST;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.NO;
-import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -12,7 +7,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
-import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCasePanelMembersExcluded;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Adjournment;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
@@ -20,7 +14,6 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.CaseDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SchedulingAndListingFields;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.ccd.presubmit.resendtogaps.ListAssistHearingMessageHelper;
 import uk.gov.hmcts.reform.sscs.service.PreviewDocumentService;
 
 @ExtendWith(MockitoExtension.class)
@@ -38,9 +31,6 @@ abstract class AdjournCaseAboutToSubmitHandlerTestBase {
 
     @Mock
     protected CaseDetails<SscsCaseData> caseDetails;
-
-    @Mock
-    protected ListAssistHearingMessageHelper hearingMessageHelper;
 
     @SuppressWarnings("unused")
     @Mock
@@ -60,29 +50,5 @@ abstract class AdjournCaseAboutToSubmitHandlerTestBase {
                 .hearingRoute(HearingRoute.GAPS)
                 .build())
             .build();
-    }
-
-    protected PreSubmitCallbackResponse<SscsCaseData> cannotBeListedAndNoDirectionsGiven() {
-        sscsCaseData.getSchedulingAndListingFields().setHearingRoute(LIST_ASSIST);
-        sscsCaseData.getAdjournment().setCanCaseBeListedRightAway(NO);
-        sscsCaseData.getAdjournment().setAreDirectionsBeingMadeToParties(NO);
-
-        return handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-    }
-
-    protected PreSubmitCallbackResponse<SscsCaseData> canBeListed() {
-        sscsCaseData.getSchedulingAndListingFields().setHearingRoute(LIST_ASSIST);
-        sscsCaseData.getAdjournment().setCanCaseBeListedRightAway(YES);
-        sscsCaseData.getAdjournment().setAreDirectionsBeingMadeToParties(NO);
-
-        return handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
-    }
-
-    protected PreSubmitCallbackResponse<SscsCaseData> cannotBeListedAndDirectionsGiven() {
-        sscsCaseData.getSchedulingAndListingFields().setHearingRoute(LIST_ASSIST);
-        sscsCaseData.getAdjournment().setCanCaseBeListedRightAway(NO);
-        sscsCaseData.getAdjournment().setAreDirectionsBeingMadeToParties(YES);
-
-        return handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
     }
 }
