+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/issueadjournment/IssueAdjournmentNoticeAboutToSubmitHandlerTest.java
@@ -3,11 +3,14 @@ package uk.gov.hmcts.reform.sscs.ccd.presubmit.issueadjournment;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.ADJOURNMENT_NOTICE;
 import static uk.gov.hmcts.reform.sscs.ccd.callback.DocumentType.DRAFT_ADJOURNMENT_NOTICE;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.HearingRoute.LIST_ASSIST;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocumentTranslationStatus.TRANSLATION_REQUIRED;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.State.HEARING;
 import static uk.gov.hmcts.reform.sscs.ccd.domain.State.NOT_LISTABLE;
@@ -18,6 +21,7 @@ import static uk.gov.hmcts.reform.sscs.ccd.domain.YesNo.YES;
 import java.time.LocalDate;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.util.ReflectionTestUtils;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
@@ -37,6 +41,10 @@ class IssueAdjournmentNoticeAboutToSubmitHandlerTest extends IssueAdjournmentNot
         when(callback.getEvent()).thenReturn(EventType.ISSUE_ADJOURNMENT_NOTICE);
         when(callback.getCaseDetails()).thenReturn(caseDetails);
         when(caseDetails.getCaseData()).thenReturn(sscsCaseData);
+
+        sscsCaseData.getAdjournment().setPreviewDocument(DocumentLink.builder()
+            .documentUrl("url")
+            .documentFilename("adjournedcasedoc.pdf").build());
     }
 
     @Test
@@ -190,4 +198,53 @@ class IssueAdjournmentNoticeAboutToSubmitHandlerTest extends IssueAdjournmentNot
             .containsOnly(DRAFT_ADJOURNMENT_NOTICE.getValue());
     }
 
+    @DisplayName("When adjournment is disabled and case is LA, then should not send any messages")
+    @Test
+    void givenFeatureFlagDisabled_thenNoMessageIsSent() {
+        ReflectionTestUtils.setField(handler, "isAdjournmentEnabled", false);
+        sscsCaseData.getSchedulingAndListingFields().setHearingRoute(LIST_ASSIST);
+
+        PreSubmitCallbackResponse<SscsCaseData> response =
+            handler.handle(ABOUT_TO_SUBMIT, callback, USER_AUTHORISATION);
+
+        verifyNoInteractions(hearingMessageHelper);
+
+        assertThat(response.getErrors()).isEmpty();
+    }
+
+    @DisplayName("When adjournment is enabled and case is LA and case cannot be listed right away "
+        + "and no directions are being made, then should send a new hearing request in hearings API")
+    @Test
+    void givenCaseCannotBeListedRightAwayAndNoDirectionsBeingMade_thenNewHearingRequestSent() {
+        PreSubmitCallbackResponse<SscsCaseData> response = cannotBeListedAndNoDirectionsGiven();
+
+        assertHearingCreatedAndAdjournmentInProgress(response);
+    }
+
+    @DisplayName("When adjournment is enabled and case is LA and case can be listed right away "
+        + "then should send a new hearing request in hearings API")
+    @Test
+    void givenCanBeListedRightAway_thenNewHearingRequestSent() {
+        PreSubmitCallbackResponse<SscsCaseData> response = canBeListed();
+
+        assertHearingCreatedAndAdjournmentInProgress(response);
+    }
+
+    private void assertHearingCreatedAndAdjournmentInProgress(PreSubmitCallbackResponse<SscsCaseData> response) {
+        verify(hearingMessageHelper, times(1))
+            .sendListAssistCreateHearingMessage(sscsCaseData.getCcdCaseId());
+
+        assertThat(response.getErrors()).isEmpty();
+    }
+
+    @DisplayName("When adjournment is enabled and case is LA and case cannot be listed right away "
+        + "and directions are being made, then should not send any messages")
+    @Test
+    void givenCaseCannotBeListedRightAwayAndDirectionsAreBeingMade_thenNoMessagesSent() {
+        PreSubmitCallbackResponse<SscsCaseData> response = cannotBeListedAndDirectionsGiven();
+
+        verifyNoInteractions(hearingMessageHelper);
+
+        assertThat(response.getErrors()).isEmpty();
+    }
 }
